refact : 모바일 컴포넌트 분리
